Option for unbounded BO

+++ src/tests/test_gp.cpp
@@ -139,12 +139,12 @@ BOOST_AUTO_TEST_CASE(test_gp_check_lf_grad)
     // Random samples and test samples
     int N = 40, M = 10;
 
-    for (size_t i = 0; i < N; i++) {
+    for (int i = 0; i < N; i++) {
         samples.push_back(tools::random_vector(4));
         observations.push_back(tools::random_vector(2));
     }
 
-    for (size_t i = 0; i < M; i++) {
+    for (int i = 0; i < M; i++) {
         test_samples.push_back(tools::random_vector(4));
         test_samples_mean.push_back(tools::random_vector(6));
         test_samples_kernel_mean.push_back(tools::random_vector(6 + 4));
@@ -156,7 +156,7 @@ BOOST_AUTO_TEST_CASE(test_gp_check_lf_grad)
 
     Eigen::VectorXd results(M);
 
-    for (size_t i = 0; i < M; i++) {
+    for (int i = 0; i < M; i++) {
         auto res = check_grad(kernel_optimization, test_samples[i], 1e-4);
         results(i) = std::get<0>(res);
         // std::cout << std::get<1>(res).transpose() << " vs " << std::get<2>(res).transpose() << " --> " << results(i) << std::endl;
@@ -166,7 +166,7 @@ BOOST_AUTO_TEST_CASE(test_gp_check_lf_grad)
 
     model::gp::KernelMeanLFOpt<Params>::KernelMeanLFOptimization<GP_t> kernel_mean_optimization(gp);
 
-    for (size_t i = 0; i < M; i++) {
+    for (int i = 0; i < M; i++) {
         auto res = check_grad(kernel_mean_optimization, test_samples_kernel_mean[i], 1e-4);
         results(i) = std::get<0>(res);
         // std::cout << std::get<1>(res).transpose() << " vs " << std::get<2>(res).transpose() << " --> " << results(i) << std::endl;
@@ -176,7 +176,7 @@ BOOST_AUTO_TEST_CASE(test_gp_check_lf_grad)
 
     model::gp::MeanLFOpt<Params>::MeanLFOptimization<GP_t> mean_optimization(gp);
 
-    for (size_t i = 0; i < M; i++) {
+    for (int i = 0; i < M; i++) {
         auto res = check_grad(mean_optimization, test_samples_mean[i], 1e-4);
         results(i) = std::get<0>(res);
         // std::cout << std::get<1>(res).transpose() << " vs " << std::get<2>(res).transpose() << " --> " << results(i) << std::endl;
